Update Impl and controller terminate

+++ ms-customer/src/main/java/com/nttdata/customer/service/Impl/CustomerServiceImpl.java
@@ -3,8 +3,8 @@ package com.nttdata.customer.service.Impl;
 import com.banking.openapi.model.CustomerRequest;
 import com.banking.openapi.model.CustomerResponse;
 import com.banking.openapi.model.ResponseDTO;
+import com.nttdata.customer.exception.types.CustomerAlreadyExistsException;
 import com.nttdata.customer.exception.types.NotFoundException;
-import com.nttdata.customer.pesistence.entity.CustomerEntity;
 import com.nttdata.customer.pesistence.repository.CustomerRepository;
 import com.nttdata.customer.service.CustomerService;
 import com.nttdata.customer.utils.AppUtils;
@@ -20,9 +20,6 @@ import reactor.core.publisher.Mono;
 @Service
 public class CustomerServiceImpl implements CustomerService {
 
-    private static final String CUSTOMER_NOT_FOUND_MESSAGE = "CustomerEntity with ID %s does not exist";
-    private static final String CUSTOMER_ALREADY_EXISTS = "CustomerEntity with Document %s already exist";
-
     @Autowired
     private CustomerRepository customerRepository;
 
@@ -35,7 +32,6 @@ public class CustomerServiceImpl implements CustomerService {
         Flux<CustomerResponse> customerDTOFlux = this.customerRepository.findAll()
                 .doOnError(error -> logger.error("Error getAllCustomer: ", error))
                 .map(AppUtils::entityToDto)
-                //.doOnNext(customerDTO -> logger.info("customer all service: {}", customerDTO))
                 .doOnTerminate(() -> logger.info("service customerRepository - end"));
 
         return Mono.just(ResponseEntity.ok(customerDTOFlux));
@@ -46,7 +42,7 @@ public class CustomerServiceImpl implements CustomerService {
         logger.info("service getCustomerById - ini");
 
         return this.customerRepository.findById(id)
-                .switchIfEmpty(Mono.error(new NotFoundException(CUSTOMER_NOT_FOUND_MESSAGE, id)))
+                .switchIfEmpty(Mono.error(new NotFoundException("CustomerEntity with ID %s does not exist", id)))
                 .doOnNext(customerEntity -> logger.info("customerEntity by id service: {}", customerEntity))
                 .flatMap(customerEntity -> Mono.just(ResponseEntity.ok(AppUtils.entityToDto(customerEntity))))
                 .doOnTerminate(() -> logger.info("service getCustomerById - end"));
@@ -55,46 +51,65 @@ public class CustomerServiceImpl implements CustomerService {
     @Override
     public Mono<ResponseEntity<CustomerResponse>> createCustomer(CustomerRequest customerRequest) {
         logger.info("service createCustomer - ini");
-//        Mono<CustomerDTO> dto = this.customerRepository.findByDocumentNumber(customerDTO.getDocumentNumber())
-//                .map(AppUtils::entityToDto)
-//                .switchIfEmpty(Mono.error(new NotFoundException(CUSTOMER_ALREADY_EXISTS, customerDTO.getDocumentNumber())));
-//        return Mono.just(ResponseEntity.ok(new CustomerDTO()));
 
-        return Mono.just(customerRequest)
-                //.doOnError(Mono.error(new IllegalArgumentException())
-                .doOnError(error -> logger.error("Error createCustomer: ", error))
-                .map(AppUtils::dtoToEntity)
-                .doOnNext(customerBefore -> logger.info("customer before create: {}", customerBefore))
-                .flatMap(this.customerRepository::insert)
-                .map(AppUtils::entityToDto)
-                .map(dto -> ResponseEntity.status(HttpStatus.CREATED).body(dto))
-                .doOnNext(customerAfter -> logger.info("customer after create: {}", customerAfter))
-                .defaultIfEmpty(ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build())
+        String documentNumber = customerRequest.getDocumentNumber();
+        return this.customerRepository.findByDocumentNumber(documentNumber)
+                .flatMap(existingCustomer -> {
+                    // Customer exist, return error.
+                    return Mono.error(new CustomerAlreadyExistsException(
+                            "Customer exist with document number: %s", customerRequest.getDocumentNumber()));
+                })
+                .switchIfEmpty(Mono.defer(() -> {
+                    // Customer not exist, return 201 created.
+                    return Mono.just(customerRequest)
+                            .map(AppUtils::dtoToEntity)
+                            .doOnNext(customerBefore -> logger.info("customer before create: {}", customerBefore))
+                            .flatMap(this.customerRepository::insert)
+                            .map(AppUtils::entityToDto)
+                            .map(dto -> ResponseEntity.status(HttpStatus.CREATED).body(dto))
+                            .doOnNext(customerAfter -> logger.info("customer after create: {}", customerAfter));
+                }))
+                .cast(ResponseEntity.class)
+                .map(responseEntity -> (ResponseEntity<CustomerResponse>) responseEntity)
                 .doOnTerminate(() -> logger.info("service createCustomer - end"));
     }
 
     @Override
-    public Mono<CustomerResponse> updateCustomerById(String id, CustomerRequest customerRequest) {
+    public Mono<ResponseEntity<CustomerResponse>> updateCustomerById(String id, CustomerRequest customerRequest) {
         logger.info("service updateCustomerById - ini");
-        // Using functional programing
+
         return this.customerRepository.findById(id)
-                .doOnError(error -> logger.error("Error updateCustomerById: ", error))
-                .flatMap(existCustomerEntity -> {
-                    CustomerEntity updatedCustomer = AppUtils.dtoToEntity(customerRequest);
-                    updatedCustomer.setId(existCustomerEntity.getId());
-                    logger.info("customer by id: {}, body customer: {}", existCustomerEntity.getId(), updatedCustomer);
-                    return this.customerRepository.save(updatedCustomer);
-                })
-                .map(AppUtils::entityToDto)
+                .switchIfEmpty(Mono.defer(() -> {
+                    // Customer not exist, return error.
+                    return Mono.error(new NotFoundException("Customer not exist with document number: %s", id));
+                }))
+                .flatMap(existCustomer -> Mono.just(customerRequest)
+                        .map(request -> {
+                            var customerEntity = AppUtils.dtoToEntity(request);
+                            // Set the ID of customer entity to updated
+                            customerEntity.setId(id);
+                            return customerEntity;
+                        })
+                        .flatMap(this.customerRepository::save)
+                        .map(AppUtils::entityToDto)
+                        .map(dto -> ResponseEntity.status(HttpStatus.OK).body(dto)))
                 .doOnTerminate(() -> logger.info("service updateCustomerById - end"));
     }
 
     @Override
-    public Mono<ResponseDTO> deleteCustomerById(String id) {
+    public Mono<ResponseEntity<ResponseDTO>> deleteCustomerById(String id) {
         logger.info("service deleteCustomerById - ini");
-        return this.customerRepository.deleteById(id)
-                .doOnError(error -> logger.error("Error deleteCustomerById: ", error))
-                .then(Mono.just(new ResponseDTO().message("CustomerEntity deleted successfully")))
-                .doOnTerminate(() -> logger.info("service deleteCustomerById - end"));
+
+        return this.customerRepository.existsById(id)
+                .flatMap(exist -> {
+                    if (exist) {
+                        return this.customerRepository.deleteById(id)
+                                .then(Mono.just(ResponseEntity.status(HttpStatus.OK)
+                                        .body(new ResponseDTO().message("Customer deleted successfully"))));
+                    } else {
+                        return Mono.error(new CustomerAlreadyExistsException(
+                                "Customer not exist with id: %s", id));
+                    }
+                });
     }
 }
